requirements/requirements.txt
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
dash==2.12.1
pefile==2023.2.7
python-magic==0.4.27
yara-python==4.3.1
pathlib
tqdm==4.65.0
joblib==1.3.1
beautifulsoup4==4.12.2
lxml==4.9.3
jupyter==1.0.0
notebook==7.0.2
PyYAML==6.0.1
colorlog==6.7.0